fix: Improve old-kernel detection
If we're falling back on the .old symlinks, there arises a situation where the
.old might point to the same kernel as the regular links. This means that
we create two entries that point to the same kernel, but are labeled differently.
This is a waste of space on the ESP and is misleading to the user.

This fixes this by checking the resolved path of both the old and new kernel
images and, if they are identical, skips backing up the old kernel.